media/gpu/VaapiVEA: Change the expectation of VideoFrame and VA Surface dimension
The original code of VaapiVEA expects the dimension of the GMB-based
VideoFrame on Encode() to be the same as the one aligned by minigbm.
This is wrong; GMB-based VideoFrame has a dimension specified by
an application (e.g. webrtc app). In other words, the size is not
aligned at all.

This fixes the wrongness. See https://tinyurl.com/frame-dimension-in-VEA
for the detail.

Bug: 982201
Test: Hangout Meet on atlas
Test: video.EncodeAccel.*
Change-Id: I7b42c9e1b2486cb7b95715185e071f0cefcf7597
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997426
Reviewed-by: David Staessens <dstaessens@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#735336}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d6a88c288ce8de8775fec58a034b5778256102ba