media/gpu/VEAtest: Cleanup memory allocation for input buffers
This code appeared wrong because it was locally allocating buffers for
the planes and then wrapping them in a VideoFrame which marked them as
unowned memory. The memory would be deallocated after the call to Encode
completes, and that posts a task so when the memory is actually accessed
later it would be invalid.  The V4L2 VEA also expects to receive owned
memory on input buffers as well.

other patches for buffer allocation unrelated to this)

Bug: None
Test: CacheLineUnalignedInputTest/** passes for I420 on trogdor (w/
Change-Id: I4552d803e5572f3f155dd984a273a1c0ffbc4b9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032050
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#737431}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a7f53f8a631730c5cdf9d19bd03d6dab14dae47c