Reland "Switch to using an explicit ended signal for HTMLMediaElement."
This is the same as the original patch, but restores the logic in
WMPI::CurrentTime() which clamps to duration upon ended. PS#1 is the
original change and all subsequent patch sets are the fixed.

Even though we have an explicit ended signal, a lot of content doesn't
have an accurate duration -- with some formats (e.g., VBR MP3, OGG) it
can't be known without a complete play-through from beginning to end.

The HTML5 spec says that upon ended, current time must equal duration.
Due to the aforementioned issue, if we rely exclusively on current
time, we can be a few milliseconds off of the duration.

[ORIGINAL DESCRIPTION BELOW]
We've long used now > dur as an implicit signal, but with poorly muxed
media this can be unreliable. Since EndedPlayback() is used for a few
important signals, we should ensure it's always right. As such, switch
to using the explicit signal known by the WebMediaPlayer.

This has been the source of a lot of flakiness over the years and we've
been thinking about doing this since 2014.

The only tricky part is in handling seeks to the end of file correctly;
our pipeline previously required that play be called before we would
trigger a real pipeline ended event. Instead we relied on HTMLME just
declaring the playback ended at now >= dur. Now we will "render" end
of stream even if playback hasn't started if we have nothing to play.

TBR=rtoy,jrummell,mlamouri,guidou

Bug: 409280, 1035472
Test: Updated unittests, all existing tests pass.
Change-Id: I46dab0049ab8145c8d647a3d8c5365a45ba84c33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033643
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#737540}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0da6f161cb3f5a83086248172c1306a61410974b