media/gpu/vaapi: fix VASurfaceID leak in VaapiVideoDecodeAccelerator
When VaapiVDA is destroyed, the Cleanup() method is called and that
calls VaapiWrapper::DestroyContextAndSurfaces() from [1]; depending
on when Cleanup() is called, |available_va_surfaces_| might not have
all VASurfaceIDs, these are never passed to vaDestroySurfaces() and
since Vaapi is opened in singleton mode, they are leaked.

The solution is to wrap them in a new self-cleaning ScopedVASurfaceID,
clarifying the ownership and guaranteeing they are destroyed. Also
the potential Vpp VaapiWrapper's Context is destroyed, which is needed
to release the VASurfaceIDs in use by it.

This only reproduces on GLK and later for H.264, and for any codec
before that, and for the "old" VDA (used still for Intel BSW, BDW,
HSW, BYT).

I am tempted to say that this was introduced in crrev.com/c/1648812,
however it landed in 77.0.3819.0 and the leak was not observed; dev
runs show that H.264 Vpp (the source of the leak) somehow seems to
hold on to these VASurfaces for too long, so perhaps sth else has
changed in M78 that caused this issue.

[1] https://source.chromium.org/chromium/chromium/src/+/master:media/gpu/vaapi/vaapi_video_decode_accelerator.cc;bpv=0;bpt=0;l=997

Bug: 1030129, 971891
Change-Id: I5639a0cb01aea482f75a18ab66317d5d7a4893b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040652
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Auto-Submit: Miguel Casas <mcasas@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#739756}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5add56eb41abd43b4482eb2eac42a4525faa8657