media/gpu/test: Group ivf frames whose timestamps are same
The served stream in webrtc can be dumped by the chrome
option [1]. When the stream structures spatial-layers, one
spatial layer is separately saved in one ivf in the dumped
ivf file.

On the other hand, to play the ivf file properly, it is needed
to group the spatial layers, which have the same timestamp in
ivf frame header. IvfFileReader [2] in webrtc does so in fact.

This applies the group logic to our own ivf parser used in
VDA test. This enables us to test SVC stream in VDA test.

[1] --force-fieldtrials=WebRTC-DecoderDataDumpDirectory/;tmp;
[2] https://source.chromium.org/chromium/chromium/src/+/master:third_party/webrtc/modules/video_coding/utility/ivf_file_reader.h

Bug: 1044816
Test: video.DecodeAccel.vp9_ksvc
Change-Id: I9a0bcfa9083dc71333ad3c727d6cbb6d902202bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032453
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#740575}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ed3b55f68a06b8f23750e581ce138156f06a35df