media/gpu/v4l2: Fix DCHECK when Initialize fails
This fixes a DCHECK that would occur if Initialize failed because then
in CreateWithOutputMode it would let the unique_ptr for the
image_processor fall out of scope and become deallocated; and we have a
sequence checker to ensure deallocation only happens on the task runner.
I would have used DeleteSoon, but that didn't seem possible because the
destructor was not public.

BUG=1050852
TEST=image_processor_test no longer DCHECKs on elm w/ 4.19 kernel

Change-Id: Ifae96c3a4206d8e94b7fdd6dd3e2ec571cc3465b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2051162
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: J Kardatzke <jkardatzke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#741148}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1e78404be990bbff3754fe8447a27406386b0e5b