media/gpu/v4l2vea: preserve crop across format update
When using imported input buffers, it is very possible that the system
will give us buffers of a coded size larger than the one we specified
when negotiating the input format with the encoder. In this case, we
have to call S_FMT again in order to set the actual input buffer format.

However doing so also has the effect of resetting the crop rectangle to
the new resolution. Make sure to set the crop right after, otherwise we
may end up with an invalid resolution in the encoded stream.

BUG=b:149451052
BUG=b:147851828
TEST=android.media.cts.EncodeDecodeTest passing.

Change-Id: I64a5f56a71b469b32d809419ff901516d1032328
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2056403
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#741443}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 26c3e2886b7308caf15bcf4c6afae090f16fcfb0