media/gpu/vaapiVDA: Fix memory leakage in profile change videos
VdaVideoDecoder holds PictureBuffer and its underlying buffer
until VdaVideoDecoder's destruction, or DismissPictureBuffer()
with the picture buffer id is called.

VaapiVideoDecodeAccelerator only calls DismissPictureBuffer()
upon a resolution event, but not on its destruction. It doesn't
matter if VdaVideoDecoder destroys immediately. Although this is
generally true, the exception is a profile change video.
VdaVideoDecoder destroys VaapiVDA and re-create VaapiVDA when the
profile in the video is changed at middle. Therefore, there may
be picture buffers that are unreleased because
DissmissPictureBuffer() is not called.

This fixes the memory leakage problem by calling
DismissPictureBuffer() in VaapiVDA's destruction.
VaapiVDA::Cleanup() is originally called in both Destroy() and
NotifyError(). This removes the call on NotifyError() to simplify
the code.

Bug: chromium:1051107
Test: No leakage in playing https://crosvideo.appspot.com/?codec=h264&loop=true&mute=true&cycle=true on atlas
Change-Id: Idb7f903a32747917dccf0d6c968f43a66678b0e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050417
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#742018}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: be8fa5df260c78cd6316e51eb763005e9344e65f