media/gpu/test/VFValidator: Get model frame by callback
VideoFrameValidator compares a video frame passed on Process()
with the model frame. The model frames are set at
Initialize() and VideoFrameValidator keeps them during the test.

The memory usage of the model frame can be large if we use
VideoFrameValidator for encoder test. We would rather give
VideoFrameValidator a callback to get a model frame, so that
we can keep model frames needed at that time in the test.

Bug: 1044452
Test: image_processor_test on kevin
Change-Id: Id3a4a76c8d149bb5e4b349ab67a087bc9bdaf45d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024769
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#744628}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cd4c2f1b85b45fd81c1a90777823237bf4bf345b