media: CreateAndInitializeVideoEncodeAcceleratorCallback to OnceCallback
This callback is used in 2 places but in different ways. The service
uses this callback as a once callback (explained below). And the
provider uses it as a repeating callback because the provider works
like a factory class. So this CL splits the callback definition into
2 separate callbacks.

A detailed explanation of why the service behaves like a once callback.

This callback is executed in an IPC handler:
* https://cs.chromium.org/chromium/src/media/mojo/services/mojo_video_encode_accelerator_service.cc?rcl=389b555af3ca96ce205c8e231429119a59aab50b&l=47

This callback is used to create an encoder, and once the class holds a
valid encoder the Initialize method can no longer be called because of
the check in:
* https://cs.chromium.org/chromium/src/media/mojo/services/mojo_video_encode_accelerator_service.cc?rcl=389b555af3ca96ce205c8e231429119a59aab50b&l=53

In theory, if the method fails it could be called repeatedly until it
succeeds. But looking at the callers it seems this doesn't happen.

The only mojo caller is in:
* https://cs.chromium.org/chromium/src/media/mojo/clients/mojo_video_encode_accelerator.cc?g=0&l=123

And that method is documented that it can only be called once:
* https://cs.chromium.org/chromium/src/media/video/video_encode_accelerator.h?g=0&l=237


This is part of the base::Callback migration.

Context: https://cs.chromium.org/chromium/src/docs/callback.md?rcl=9fcc3764aea8f97e9f6de4a9ee61d554e67edcda&l=40

Bug: 714018
Change-Id: I1f51ff76b679c891740d2848e9d4c573ef315db0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2069329
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Jose Lopes <jabolopes@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#744977}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31d25be795886ffe1ee00514081c349bf85247a8