[Audio Process] Enable Kill-on-Hang by default on Linux.
There has been a recent increase in occasions where the audio process
has been hanging on Linux. To remedy the issue we are launching the
Kill-on-Hang feature by default on Linux. The feature has been running on
Windows and Mac for the past several months, and proved to be effective
in these situations.

Bug: 1043040
Change-Id: I2cb79a518faabc08a7c9b02890e9042abfd67c06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2081818
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Armando Miraglia <armax@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#745896}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 71825939c432c440fa53ef4016372076e2c6114a