media/gpu/V4L2VDA: Fix sequence check failing in device poller on shutdown.
This CL fixes a sequence check failing when destroying the device
poller. This can happen when a video decoder is destroyed before
initialization, causing destruction to not be performed on the video
decoder thread as it was never started.

TEST=./video_decode_accelerator_tests on hana

BUG=b:150650090

Change-Id: I3f014c70cc5e7ffdd54d706a93ad12e014189a88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2084672
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748156}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1e3ecfea79f472f33b47c985f98cf21a51f2a7b7