media/gpu/v4l2IP: Fix the wrong assignment to plane length in v4l2 DMABUF buffer
V4L2IP sets plane length to the area of coded size and offset.
This is wrong if the width of coded_size is different from the
stride of the underlying buffer. This CL corrects it by setting
to the size of GpuMemoryBufferPlane.

Bug: None
Test: ip_test --gtest_filter=NV12* on kukui
Change-Id: I3e49e6ecd2442e00b8d6705cfb3e333efad0a9f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2096427
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749539}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fce448fa6889731026a3715542155512131302f3