[fuchsia] Use FIDL TestBase for TestAudioCapturer.
Using the AudioCapturer_TestBase allows the TestAudioCapturer to compile
even if new APIs are added to the AudioCapturer API, and reduces the
amount of boilerplate for APIs which are not expected to be called.

This is required in order to unblock the next Fuchsia SDK roll.

TBR: sergeyu
Bug: fuchsia:48420, 1047390
Change-Id: I15ccc2c50bba2ea6bd30d2508b5d7e6e668d73aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106582
Auto-Submit: Wez <wez@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: Wez <wez@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750968}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3188eea3fb369da894a1f82c872fe9978a26f078