media/gpu/vaapi: Filter out unused VAEntrypoints in GetEntryPointsForProfile()
VaapiWrapper queries capabilities with all the supported
VAEntrypoints for the VAProfile. VA functions with some
VAEntrypoint fails. It results in the output in chorme log and
also chrome://gpu as ERROR. The failure is red herring but
sufficient to confuse developers and users.
To avoid this, we filter out VAEntrypoints upon querying
capabilities that are not used in chrome scenario.

Bug: 1057196
Test: vaapi_unittest on eve
Change-Id: I1160fde180e53a08125dda6b6645152636cbda4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2108305
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752728}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 17b570ab6ccf5e7d3c5b9f04bcb9b9463738b0ae