Adds NotifyError for Status types
And, as a use case, logs failure reason for decoders when they
can't be initialized.

This currently just dumps the status as JSON into the media log. I've
not implemented a frontend for this in devtools yet.

Change-Id: I3aea5f027078e3f2979eaeeebc08549aa8a393a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2141622
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Ted Meyer <tmathmeyer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#757508}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c1dab74f444a18c09c6dc1179d31db8758a21603